asm-350 - move stillInvolved validators to seperate file
